image-list: Add signal `image-added`
Similar to what was described in the previous commit, we need this to
later map a container to its image. Since we don't know if the image or
the container will be loaded first, we need these signals on both list
types. In case the container came before the image, we can search all
containers when the image arrives.