fix(container-creation-page): Make memory limit work again
The new models in `posman-api-rs` have made most of the fields in
`LinuxResources` mandatory. This leads to errors if only individual
limits are to be set selectively. Podman then tries to set the other
limits as well. This in turn can lead to errors when starting
containers.

All fields are therefore made optional again.

Furthermore, the megabytes are now correctly converted to bytes.