refactor(application): Use AdwApplication as base
We change the base class of the application to `AdwApplication` to
simplify the styling of the app.

In this turn, this commit also removes the ability to recolor the header
bar. This will eventually come back when there will be a libadwaita
recoloring API. But at the moment, this is to hacky and the burden to
maintain it is not worth it.