feat(inspection-page): Load inspection data during construction
Previously, the inspection data was loaded on the corresponding detail
page after pressing the button and then passed to the inspection page.
But this could lead to short delays.

Therefore, in this commit, the loading of the inspection data is moved
to the inspection page. This way it is possible to switch to it
immediately. Until the data is loaded, a spinner is displayed.