fix(container-list): Call `items_changed` immediately
After a container has been removed from the baking list, we should call
`items_changed` immediately after that.

Before this commit, we would call `container_removed` first. This
triggered an update of the properties filter in the
`PdsContainersGroup`. In this way, the view and model got out of sync
resulting in wrong container row filtering.